Refactoring/mypage managerpage

+++ src/components/Feed/Category/CategoryBox.tsx
@@ -1,19 +1,23 @@
-import React, { useContext, useState } from 'react';
+import React, { useState } from 'react';
 import * as UI from './style';
 import AddTodo from '../Todo/AddTodo';
 import TodoBox from '../Todo/TodoBox';
-import { Category } from '../interfaces';
+import { Category, PatchFeedPayload } from '../interfaces';
 import useInput from '../../../hooks/common/useInput';
 
 import { BsX } from '@react-icons/all-files/bs/BsX';
 import { useDeleteFeed } from '../../../api/hooks/Feed/useDeleteFeed';
 import { recoilTabState } from '../../../states/recoilTabState';
 import { useRecoilValue } from 'recoil';
+import Swal from 'sweetalert2';
+import { COLOR } from '../../../styles/colors';
+import { usePatchFeed } from '../../../api/hooks/Feed/usePatchFeed';
 
 const CategoryBox = ({ categoryId, categoryName, todos }: Category) => {
   const [openTodoInput, setOpenTodoInput] = useState<boolean>(false);
   const [AddTodoState, setAddTodoHandler, setAddTodoState] = useInput(15);
   const tab = useRecoilValue(recoilTabState);
+  const { mutate } = usePatchFeed();
 
   // category 내부의 + 버튼 눌렀을 때의 function
   const TodoPlusHandler = () => {
@@ -45,14 +49,66 @@ const CategoryBox = ({ categoryId, categoryName, todos }: Category) => {
 
   const { deleteFeed } = useDeleteFeed();
 
+  // 삭제버튼 클릭 시 더블체크
   const deleteBtnHandler = () => {
-    deleteFeed({ type: 'category', id: categoryId });
+    Swal.fire({
+      title: '카테고리를 삭제하시겠어요?',
+      text: '하위 todo도 삭제됩니다.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: 'black',
+      cancelButtonColor: 'gray',
+      confirmButtonText: '네.',
+      cancelButtonText: '아니요.',
+      reverseButtons: true,
+      customClass: {
+        title: 'sweet-alert-title',
+        htmlContainer: 'sweet-alert-html-container',
+      },
+      didOpen: () => {
+        const titleSpan = document.querySelector('.sweet-alert-title');
+        const textSpan = document.querySelector('.sweet-alert-html-container');
+        if (titleSpan && textSpan) {
+          (titleSpan as HTMLElement).style.fontSize = '18px';
+          (titleSpan as HTMLElement).style.fontWeight = 'bold';
+          (textSpan as HTMLElement).style.color = 'gray';
+        }
+      },
+    }).then(result => {
+      if (result.isConfirmed) {
+        deleteFeed({ type: 'category', id: categoryId });
+      }
+    });
+  };
+
+  // 카테고리 수정기능
+  const ModifyCategory = async () => {
+    const { value: inputValue } = await Swal.fire({
+      title: '카테고리를 수정하시겠어요?',
+      input: 'text',
+      showCancelButton: true,
+      confirmButtonColor: COLOR.PAGE_BLUE,
+      cancelButtonColor: COLOR.VACATION_RED,
+      confirmButtonText: '입력한 내용으로 바꿀래요.',
+      cancelButtonText: '아니요, 안바꿀래요.',
+    });
+
+    if (inputValue) {
+      const payload: PatchFeedPayload = {
+        feed: 'category',
+        id: categoryId,
+        content: {
+          category: inputValue,
+        },
+      };
+      mutate(payload);
+    }
   };
 
   return (
     <>
       <UI.StCategoryWrapper>
-        <UI.StCategoryTitleBlock tab={tab}>
+        <UI.StCategoryTitleBlock tab={tab} onClick={ModifyCategory}>
           <UI.StCircleBlock />
           <UI.StCategoryH3>{categoryName}</UI.StCategoryH3>
         </UI.StCategoryTitleBlock>
